+++ packages/moonstone/CHANGELOG.md
@@ -21,7 +21,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/Divider` to now behave as a simple horizontal line when no text content is provided
 - `moonstone/Scrollable` to not display scrollbar controls by default.
 - `moonstone/DatePicker` and `moonstone/TimePicker` to emit `onChange` event whenever the value is changed, not just when the component is closed.
-- `moonstone/Button` to now be wrapped in `ui/Holdable`.
+- `moonstone/IncrementSliderButton` to now be wrapped in `ui/Holdable`.
 
 ### Removed